Reorg IndexReader to separate files
